Reduce duplication in infinite scroll hooks
