+++ src/common/dataHooks/useSearchAllSitesLoader.js
@@ -1,43 +1,26 @@
-import { useInfiniteQuery } from '@tanstack/react-query'
 import PropTypes from 'prop-types'
 
 // FPCC
-import useLoader from 'common/hooks/useLoader'
+import useInfiniteScroll from 'common/dataHooks/useInfiniteScroll'
 import api from 'services/api'
 import { SEARCH } from 'common/constants'
-import { searchResponseAdaptor } from 'common/dataAdaptors'
+import { searchResultAdaptor } from 'common/dataAdaptors'
 
-/**
- * Calls search API and provides search results and infinite scroll info.
- */
 function useSearchAllSitesLoader({ enabled, searchParams }) {
   const searchParamString = searchParams.toString()
 
-  // Fetch search results
-  const response = useInfiniteQuery({
+  const infiniteQueryResponse = useInfiniteScroll({
     queryKey: [SEARCH, 'cross-site', searchParamString],
     queryFn: ({ pageParam = 1 }) =>
       api.search.getFVWideSearch({
         searchParams: searchParamString,
         pageParam,
       }),
-    getNextPageParam: (currentPage) => currentPage.next,
     enabled: !!enabled,
-    refetchOnWindowFocus: false,
-    refetchOnReconnect: false,
-    select: (responseData) => ({
-      pages: searchResponseAdaptor(responseData),
-      pageParams: responseData.pageParams,
-    }),
+    resultAdaptor: searchResultAdaptor,
   })
 
-  const { infiniteScroll, loadRef } = useLoader({ response })
-
-  return {
-    ...response,
-    infiniteScroll,
-    loadRef,
-  }
+  return infiniteQueryResponse
 }
 
 // PROPTYPES
